ajout triangle + test plus propre

+++ neo4j-java-bolt/src/main/java/example/movies/backend/DbRoutes.java
@@ -6,17 +6,22 @@ import spark.servlet.SparkApplication;
 
 import javax.servlet.MultipartConfigElement;
 import javax.servlet.http.Part;
+import java.io.File;
 import java.io.InputStream;
 import java.net.URLDecoder;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
+import java.nio.file.StandardCopyOption;
 import java.util.Collection;
+import java.util.Map;
+
 
 import static spark.Spark.get;
 import static spark.Spark.post ;
 import static spark.Spark.path;
+
 
 public class DbRoutes implements SparkApplication {
 
@@ -31,39 +36,44 @@ public class DbRoutes implements SparkApplication {
     public void init() {
 
         post("/test", (request, response) -> {
-            String location = "image";          // the directory location where files will be stored
+            String location = "/home/denis/5A/"; // the directory location where files will be stored
+            String answer= "OK";
             long maxFileSize = 100000000;       // the maximum size allowed for uploaded files
             long maxRequestSize = 100000000;    // the maximum size allowed for multipart/form-data requests
             int fileSizeThreshold = 1024;       // the size threshold after which files will be written to disk
 
             MultipartConfigElement multipartConfigElement = new MultipartConfigElement(
                     location, maxFileSize, maxRequestSize, fileSizeThreshold);
-            request.raw().setAttribute("org.apache.tomcatmultipartConfig",
+            request.raw().setAttribute("org.eclipse.jetty.multipartConfig",
                     multipartConfigElement);
 
             Collection<Part> parts = request.raw().getParts();
+            /*
             for (Part part : parts) {
                 System.out.println("Name: " + part.getName());
                 System.out.println("Size: " + part.getSize());
                 System.out.println("Filename: " + part.getSubmittedFileName());
             }
+             */
 
             String fName = request.raw().getPart("file").getSubmittedFileName();
-            System.out.println("Title: " + request.raw().getParameter("title"));
             System.out.println("File: " + fName);
 
             Part uploadedFile = request.raw().getPart("file");
-            Path out = Paths.get("image/" + fName);
-            try (final InputStream in = uploadedFile.getInputStream()) {
-                Files.copy(in, out);
-                uploadedFile.delete();
-            }
-// cleanup
-            multipartConfigElement = null;
-            parts = null;
-            uploadedFile = null;
 
-            return "OK";
+            File dest = new File("/home/denis/5A/testUpload.txt");
+            if(dest.createNewFile() ) {
+                Path out = dest.toPath() ;
+                //Path out = Paths.get("home/denis/5A/" + fName+ "Server");
+                try (final InputStream in = uploadedFile.getInputStream()) {
+                    Files.copy(in, out, StandardCopyOption.REPLACE_EXISTING);
+                    uploadedFile.delete();
+                }
+                //service.importCSV("/home/denis/5A/testUpload.txt", " ", true) ;
+            }
+            else
+                answer = "NOK" ;
+            return answer;
         }) ;
 
         //return a node in the graph and all his neighboors.
@@ -115,6 +125,18 @@ public class DbRoutes implements SparkApplication {
                 if(!service.graph_exists(name))
                     service.create_graph(name, "Person", "Send", directed);
                 return gson.toJson(service.labelPropagation(name, "stream")) ;
+            }) ;
+
+            get("/triangle/:name", (req,res)-> {
+                String name = URLDecoder.decode(req.params("name"),  StandardCharsets.UTF_8) ;
+                Object result ;
+                if(service.graph_exists(name)) {
+                     result = gson.toJson(service.triangle(name, "stream")) ;
+                }
+                else {
+                    result = "Graph does not exists" ;
+                }
+                return result ;
             }) ;
         });
